Put vault creation in playtrough.sh, also minor changes
General:
playtrough.sh step 1 - obtaining password from keyboard or from an
exported environment variable.
playtrough.sh step 2 - getting access to the vault, opening and even
creating it if necessary.

Changes by file:
.flask-ci-cd-project-root-marker - created a file in project root to
test for - a simple way to make sure the script was executed in the
project root as expected.

vault_management.sh:
1. Made the script set ownership of the vault mountpoint after
mounting the vault to the user executing the script.
2. Made the script only inform about sudo password requirement if it
hasn't been cached

playtrough.sh:
1. Added variables needed for vault access/creation
2. Modified usage() to use two arguments instead of one containing a
dash if we want to submit the last step number. Fishing out arguments
from inside a dash can be done, but it's unnecessary work.
3. Added a die() function, that is used to exit the script with an
error. In addition it reports in which step the error occurred.
4. Added a check for project root marker file and error if it's not
present. The rest of the script assumes we are in the project root,
it's wise to make sure.
5. Used exported VAULT_PASSPHRASE instead of
prompting the user for it, if it's available.
6. Told the user that sudo password might be required to
open/mount/create LUKS container.
7. Also warned the user that it takes some time to create/open LUKS
containers.
8. Step 2 will do what's needed to establish vault access and inform
about it or fail if it couldn't do it.
9. Reworded step 3 description to state that an existing ssh key will
not be overwritten.
10. Fixed an 'a/the' typo in step 13 description.