Fix Terraform error 'SSH key already uploaded' with a data resource
If the SSH key has already been uploaded and some time passed since it
happened, there will/can be an error. Terraform refuses to upload a
key when another key with the same name already exists.

It can be solved by creating a data structure "aws_key_pair" with the
name we want to use. And then, when we try to upload the key, creating
a resource structure, we make sure to do it only if there is no key
existing already.

To achieve conditional creation of resources (not supported in
Terraform) we use a trick - 'count' keyword that allows us to create a
certain number of items. For the value of the 'count' we give a
conditional expression, `data.aws_key_pair.existing != null ? 0 : 1`.
This expression returns 0 if there is already a key, which in turns
stops Terraform from trying to re-upload the key and saves us from the
error.