Make Terraform add IP to inventory on each run of 'terraform apply'
Terraform wouldn't normally execute the local script if the instance
was already created. This was an edge case that came out when testing
and I found out how to have Terraform do something each time it was
run. It required using a hack in form of a condition with value of
"timestamp()". Using a boolean wouldn't work, because Terraform checks
not for the condition to be true, but for any changes made to the
values. Timestamp is guaranteed to have a different value each second,
which is good enough for us.

It shows that Terraform wasn't really made for things like that and it
probably should've been done with bash/python used to run Terraform,
and not Terraform itself.