Refactor the plan for the project to help ublock the progress
From the code journal:
Looks like what caused the problem was wrong order of tasks. You can't
configure Jenkins automatically if you don't know what to configure
for. For that you need to have a finished and working Jenkins
pipeline. For that you need the code repo to contain the code to
manually generate the artifact. So the proper order would be:
1. Code repo has all it needs to produce a working artifact, in this
   case that would most likely be a Dockerfile.
2. Code repo contains a Jenkinsfile which is just the automated
   creation of the artifact mentioned in previous step (normally it
   would also include testing, before creating the artifact and after,
   but this would add too much complexity).
3. Jenkins installation can be configured to create the pipeline for
   the code-repo based on a Jenkinsfile in the repo.