Change the repo to more specific so it can work with Alma Linux
Because Docker's repo doesn't yet aknowledge AlmaLinux, we need to use
a dirty trick and add a CentOS repo. But because adding it in a casual
way wouldn't work, we need to be more specific. Luckily, dnf on our
machine will recognize the $basearch and $releasever parameters, so we
don't have to be 100% specific with the path.