Add aws (cli) installation to the playbook.yml (Jenkins Server)
Installs aws(cli) only if it's not present.
Used a playbook 'block' and 'when' coupled with a registered variable.
Also changed all module names in the playbook to use the fully
qualified module name (because the code for aws-cli installation
already had it and it was quick to do with an Emacs macro).