Proofread README and fix typos tHroughout the project
Most common mistakes were:
- trough instead of through - about 50 occurrences
- it's instead of its (and similar) - about 15 occurrences

The general term for the mistake of writing "it's" instead of "its" is
called a possessive vs. contraction error. This error occurs when the
contraction "it's" (meaning "it is" or "it has") is mistakenly used in
place of the possessive pronoun "its" (indicating possession). This is
a common grammatical mistake.