Update - Finished the step of uploading the docker image to ECR
Finished the step of manually uploading the Docker image to the ECR
repo. I re-did the manual upload at the end of the "Upload Docker
image to ECR" task, because a lot of time has passed between it has
been started.

Feels good. A good practice that helped here was making often used but
complicated shell commands into scripts and putting them in the
project's bin/ directory.