Add pre-commit git hook to export org to markdown
This pre-commit hook will export two documentation files from org to
markdown. It also hides IP addresses from the hosts.ini file (that
belongs to Ansible).

The hook should be easy to re-use or to expand to more files, because
I used variables and a for loop.

It might require changing the export method, because with two files it
already takes a moment on an older PC. Might use an emacs-lisp script,
which will just accept the filenames as parameters. I suspect that the
delay comes from starting and closing emacs twice and not the
exporting itself.