Save main.tf and vars.tf before trying to fix ssh key handling
For now, the code that was supposed to detect an ssh key already
uploaded doesn't work, as it creates an error. Trying to fix it and
make it more robust, so that it can:
1. detect an existing key and verify it's the same as the local key
2. re-upload the key, if there is none or the existing is different.

It's possible it will be to difficult, in that case I will revert here
and try something easier, like just uploading the key once and having
the user fix stuff manually if something is wrong (remote key being
different than what we have).