Add first version of the script to compare remote SSH key with local
Had an issue with SSH key not working. It was caused by Terraform not
uploading an SSH key to aws, because there already was a key with the
given name uploaded. The keys were not the same, but had the same
name.

This script calculates key fingerprints and compares them.