Replace complicated SSH logic with a simple aws_key_pair resource
I believed in ChatGPT's logic thinking that a 'data' type resource can
be conditionally created only if an resource with given name
exists. That's not true, such an attempt will cause an error. It's a
rare occurrence and therefore there's no point in writing scripts
around checking the ssh keys. I'm not even sure if ssh keys should be
managed with Terraform like that.